Refactor await method
