+++ src/main/groovy/io/seqera/wave/service/blob/BlobStore.groovy
@@ -1,10 +1,6 @@
 package io.seqera.wave.service.blob
 
 import java.time.Duration
-import java.util.concurrent.CompletableFuture
-
-import groovy.transform.CompileStatic
-
 /**
  * Implement a distributed store for blob cache entry.
  *
@@ -34,49 +30,4 @@ interface BlobStore {
      */
     boolean storeIfAbsent(String key, BlobCacheInfo info)
 
-    /**
-     * Await for the container layer blob download
-     *
-     * @param key
-     *      The container blob unique key
-     * @return
-     *      the {@link CompletableFuture} holding the {@link BlobCacheInfo} associated with
-     *      specified blob key or {@code null} if no blob record is associated for the
-     *      given key
-     */
-    default BlobCacheInfo awaitCacheStore(String key) {
-        final result = getBlob(key)
-        return result ? Waiter.awaitCompletion(this,key,result) : null
-    }
-
-    /**
-     * Implement waiter common logic
-     */
-    @CompileStatic
-    private static class Waiter {
-
-        static BlobCacheInfo awaitCompletion(BlobStore store, String key, BlobCacheInfo current) {
-            final beg = System.currentTimeMillis()
-            // add 10% delay gap to prevent race condition with timeout expiration
-            final max = (store.timeout.toMillis() * 1.10) as long
-            while( true ) {
-                if( current==null ) {
-                    return BlobCacheInfo.unknown()
-                }
-
-                // check is completed
-                if( current.done() ) {
-                    return current
-                }
-                // check if it's timed out
-                final delta = System.currentTimeMillis()-beg
-                if( delta > max )
-                    throw new TransferTimeoutException("Blob download '$key' timed out")
-                // sleep a bit
-                Thread.sleep(store.delay.toMillis())
-                // fetch the build status again
-                current = store.getBlob(key)
-            }
-        }
-    }
 }
